chore: move types in d.ts file and enum in their own
