+++ .eslintrc.json
@@ -28,7 +28,7 @@
     "no-lonely-if": "error",
     "no-multi-spaces": "error",
     "no-multiple-empty-lines": ["error", { "max": 2, "maxEOF": 1, "maxBOF": 0 }],
-    "no-shadow": ["error", { "allow": ["err", "resolve", "reject"] }],
+    "@typescript-eslint/no-shadow": ["error", { "allow": ["err", "resolve", "reject"] }],
     "no-trailing-spaces": ["error"],
     "no-var": "error",
     "object-curly-spacing": ["error", "always"],
